Updated helper method

+++ src/main/java/se/Matryoshika/Underworld/Utils/Print.java
@@ -4,7 +4,7 @@ import net.minecraft.util.math.BlockPos;
 
 public class Print {
 	
-	public static String print(int x, int y, int z){
+	public static String print(double x, double y, double z){
 		return x+", "+y+", "+z;
 	}
